biscuit asm fixes
I'll be damned if I'm 100% sure, but it seems that the assertion of unit width only applies to the epistate number (1 or 2) and the Waston/Crick strand (+/-), whereas cg and snp are fields 5 and 7 (4 and 6 in 0-based indexing) are the correct ones for that.  At least, I think that's what's supposed to happen here.  It does compile and run without issue, although with the epireads I feed it, no output is produced.