Change epiread format to BED format and RLE strings
For now, S(hut) / U(nmethylated) / M(ethylated) / O(pen) letter always
falls on the C in the CG or GC context. Insertions are included as lower
case letters in the RLE string. Deletions are included as "D" and soft
clipped bases are included as "P". Old epiread format can be accessed
with -O argument. This will be deprecated in the future.