Fix crash inducing bug, incorrect region end in pileup/epiread, loss of reads with region
The crash inducing bug was caused by not providing enough memory for encoded RLE string.
When the 1's are removed from the encoded string, I can reduce the memory back to being
equal to the length of the unencoded string, but for now, I have to ask for twice the
length of the unencoded string.

Previously, the parsed end of the region given to pileup and epiread was being incremented
by 1. This only needed to be done for the parsed beginning, as the beginning was 0-based,
while the end is 1-based. This lines biscuit up better with how samtools does the region
parsing.

Reads were getting lost when providing a region as the logic to avoid double counting reads
was doing exactly what it should. However, when given a region, the first and last windows
may have data that starts before or after the end of the specified region (which is correct
behavior). This was fixed in this commit.