Fix implicit fallthrough warnings
The fallthroughs would never occur as the wzfatal function causes an error
and the program would return. However, the compiler wouldn't recognize
this and would throw a warning. Fixed by adding a break statement that
the code never reaches. This is probably a band-aid fix that doesn't need
to be there, but probably implies that the code could use some slight
refactoring to handle this all properly. Oh well, it works, there are
no more warnings, and it can get dealt with later, if at all.