Add long read functionality for epiread
This was done in a way that works, but in the future it would be good
to refactor the code slightly to accept a CLI option to set the variables
that were added/modified by this commit. It will provide more flexibility
for the user and hopefully, better memory management. Also, if there
isn't a check for read length already, then make sure this exists so that
the code will fail gracefully, instead of providing a useless seg fault.