Hot fix for old epiread format
Adds a "-" for CGs that were filtered for low base quality, position in
read (start or end bases), or double counting overlapped r1 and r2. One
further case that needs to be addressed in the future is how to handle
SNPs, as the CG position is always relative to the C, but SNPs can occur
in the G, so need to figure out how to handle that instance.