Address Issue #25 for handling overlapping reads
Read length is now determined for mate read using the MC (mate cigar)
auxiliary tag. If MC is missing, then read lengths are assumed to be the
same. While the length of the mate read could have been added, there are
assorted tools that can easily add the MC tag, which makes using already
aligned data easier. I'm not aware of a tool that can easily add the mate
read length as an auxiliary tag. This also necessitated a change to the
alignment output, where biscuit will now natively output the MC/MQ mate
tags during alignment (handled in a previous commit).

For overlapping reads, it doesn't matter if the reads are marked as proper
pairs or not. By default improper pairs will be filtered before reaching
the place to check for overlap. If the user requests that improper pairs
be included, then it would make sense to treat overlapping reads the same,
regardless of the "properness" of the pairing.