Hide no filter secondaries in biscuit epiread
Allowing secondary reads in the epiBED/epiread files doesn't make
much sense. They mess with correct methylation level calculations
and being able to properly collapse to the fragment level. The
option will still remain under-the-hood, but it will be those who
have the specific use-case (as yet unknown) for it.