Handle read comments in a more intuitive manner
Comments were previously added in the SAM tags section of the SAM entry.
However, they weren't placed in a tag, so they weren't necessarily the
easiest to pull out. Now, comments are tacked on to the read name (with
an underscore used as the connector). I believe this is how Bismark does
it, and it should end up being easier to now extract information (like
barcodes).