Update htslib to version 1.18
After previously removing the submodule, htslib now is available
directly during the git clone (without needing the --recursive option).
This should hopefully make transitioning to cmake easier, but that will
remain to be seen. Tests (in both htslib and htscodecs) and samples have
been left in as these files are released with the library; however, if a
smaller binary is desired, it's probable these could be removed without
loss of functionality.