Resolve using possibly uninit'ed value in epiread
While this resolves the problem of using a possibly uninitialized value
in an if-statement, it will possibily changes the output of the old
epiread format for reads with no CGs in them. This does fix undefined
behavior, but I think a better solution would solve both this and retain
original behavior.  That said, this functionality is maintained for
legacy purposes and should be deprecated/removed going in the next major
release, whenever that happens.