resolve edge case bug
when a the mapped start position (from SAM) starts in one region, but
the start adjusted for soft clipping moves it out of the region, the
read won't be printed (since it is only found in the one region, but not
the previous). Use the mapped start position to determine which region
to print a read from.