Fix writing to inaccessible memory address
Just checking the read length wasn't enough to catch writing to out of
bounds locations. Also needed to check that the length of the read plus
the number of deletions wouldn't be too long for the buffer. Also moved
the buffer out of the read loop, so now each thread gets its own buffer
that will be overwritten with 0s in the read loop.